Move entity structs to common place for public access
