+++ plugins/processors/awsentity/processor_test.go
@@ -6,6 +6,7 @@ package awsentity
 import (
 	"bytes"
 	"context"
+	"github.com/aws/amazon-cloudwatch-agent/internal/entity"
 	"os"
 	"testing"
 
@@ -19,7 +20,6 @@ import (
 	"github.com/aws/amazon-cloudwatch-agent/extension/entitystore"
 	"github.com/aws/amazon-cloudwatch-agent/internal/k8sCommon/k8sclient"
 	"github.com/aws/amazon-cloudwatch-agent/plugins/processors/awsentity/entityattributes"
-	"github.com/aws/amazon-cloudwatch-agent/plugins/processors/awsentity/internal/entitytransformer"
 	"github.com/aws/amazon-cloudwatch-agent/translator/config"
 )
 
@@ -569,14 +569,14 @@ func TestProcessMetricsResourceAttributeScraping(t *testing.T) {
 			},
 			config: &Config{
 				EntityType: attributeService,
-				TransformEntity: &entitytransformer.EntityTransform{
-					KeyAttributes: []entitytransformer.KeyPair{
+				TransformEntity: &entity.EntityTransform{
+					KeyAttributes: []entity.KeyPair{
 						{
 							Key:   "Name",
 							Value: "override-service",
 						},
 					},
-					Attributes: []entitytransformer.KeyPair{
+					Attributes: []entity.KeyPair{
 						{
 							Key:   "AWS.ServiceNameSource",
 							Value: "UserConfiguration",
@@ -604,14 +604,14 @@ func TestProcessMetricsResourceAttributeScraping(t *testing.T) {
 			},
 			config: &Config{
 				EntityType: attributeService,
-				TransformEntity: &entitytransformer.EntityTransform{
-					KeyAttributes: []entitytransformer.KeyPair{
+				TransformEntity: &entity.EntityTransform{
+					KeyAttributes: []entity.KeyPair{
 						{
 							Key:   "Name",
 							Value: "override-service",
 						},
 					},
-					Attributes: []entitytransformer.KeyPair{
+					Attributes: []entity.KeyPair{
 						{
 							Key:   "AWS.ServiceNameSource",
 							Value: "UserConfiguration",
